# Speed up ast.tokenLocation